requirements.txt
Flask
flask
flask-cors
psycopg2-binary
Werkzeug==2.3.2
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
python-dotenv
